initial fetch bug fixed. Permission request bug fixed

+++ app/src/main/java/com/onbiron/forecastmvvm/data/network/response/future/Daily.kt
@@ -7,7 +7,7 @@ data class Daily(
     val clouds: Double,
     @SerializedName("dew_point")
     val dewPoint: Double,
-    val dt: Double,
+    val dt: Long,
     @SerializedName("feels_like")
     val feelsLike: FeelsLike,
     val humidity: Double,
